requirements.txt
torch
pandas
scikit-learn
kafka-python
openai
